pyproject.toml
[tool.poetry]
name = "basic-data-platform"
version = "0.1.0"
description = ""
authors = ["ikarabulut <i.karabulut93@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
pytest = "^7.4.3"
fastapi = "^0.104.1"
httpx = "^0.25.1"
pydantic = "^2.4.2"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
